requirements.txt
Django==4.0.0
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2==2.9.7
